requirements.txt
streamlit>=1.35.0
plotly>=5.18.0
pandas>=2.0.0
anthropic>=0.25.0
playwright>=1.49.0
python-dotenv>=1.0.0
requests>=2.31.0
psycopg2-binary>=2.9.0
beautifulsoup4>=4.12.0
